backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
datapizza-ai==0.0.9
datapizza-ai-clients-openai==0.0.9
datapizza-ai-core==0.0.10
datapizza-ai-embedders-openai==0.0.5
datapizza-ai-vectorstores-qdrant==0.0.4
distro==1.9.0
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.21.1
mdurl==0.1.2
numpy==2.3.4
openai==2.8.0
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
portalocker==3.2.0
protobuf==6.33.1
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
qdrant-client==1.15.1
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sse-starlette==3.0.3
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
Werkzeug==3.1.3
zipp==3.23.0
